move all fetches to /lib/server-fetches.ts
