+++ frontend/src/lib/server-fetches.ts
@@ -1,7 +1,15 @@
-import { BThread } from "@/lib/types";
 import { auth } from "@/lib/auth";
 import { getSettings } from "@/lib/cookies-server";
 import { fetcher } from "@/lib/fetcher";
+import {
+  BThread,
+  BMessage,
+  BToolMetadataDetailed,
+  BToolMetadata,
+  CustomError,
+  ToolDetailedMetadata,
+  ToolMetadata,
+} from "@/lib/types";
 
 export async function getThread(threadId: string): Promise<BThread | null> {
   try {
@@ -34,6 +42,54 @@ export async function getThread(threadId: string): Promise<BThread | null> {
   }
 }
 
+export async function getMessages(
+  threadId: string,
+): Promise<BMessage[] | null> {
+  const session = await auth();
+  if (!session?.accessToken) {
+    throw new Error("No session found");
+  }
+
+  try {
+    const response = await fetcher({
+      path: "/threads/{threadId}/messages",
+      pathParams: { threadId },
+      headers: { Authorization: `Bearer ${session.accessToken}` },
+      next: {
+        tags: [`thread/${threadId}/messages`],
+      },
+    });
+
+    return response as Promise<BMessage[]>;
+  } catch (error) {
+    if ((error as CustomError).statusCode === 404) {
+      return [];
+    } else {
+      throw error;
+    }
+  }
+}
+
+export async function getTools(): Promise<ToolMetadata[]> {
+  const session = await auth();
+  if (!session?.accessToken) {
+    throw new Error("No session found");
+  }
+
+  const tools = (await fetcher({
+    method: "GET",
+    path: "/tools",
+    headers: { Authorization: `Bearer ${session.accessToken}` },
+  })) as BToolMetadata[];
+
+  return tools
+    .map((tool) => ({
+      name: tool.name,
+      nameFrontend: tool.name_frontend,
+    }))
+    .sort((a, b) => a.nameFrontend.localeCompare(b.nameFrontend));
+}
+
 export async function getToolList() {
   const session = await auth();
   if (!session?.accessToken) {
@@ -50,4 +106,38 @@ export async function getToolList() {
       return { slug: tool.name, label: tool.name_frontend };
     })
     .sort((a, b) => a.label.localeCompare(b.label));
+}
+
+export async function getTool(
+  toolName: string,
+): Promise<ToolDetailedMetadata | null> {
+  const session = await auth();
+  if (!session?.accessToken) {
+    throw new Error("No session found");
+  }
+
+  try {
+    const tool = (await fetcher({
+      method: "GET",
+      path: "/tools/{name}",
+      pathParams: { name: toolName },
+      headers: { Authorization: `Bearer ${session.accessToken}` },
+    })) as BToolMetadataDetailed;
+
+    return {
+      name: tool.name,
+      nameFrontend: tool.name_frontend,
+      description: tool.description,
+      descriptionFrontend: tool.description_frontend,
+      inputSchema: tool.input_schema,
+      hil: tool.hil,
+      isOnline: tool.is_online,
+    };
+  } catch (error) {
+    if ((error as CustomError).statusCode === 404) {
+      return null;
+    } else {
+      throw error;
+    }
+  }
 }
